requirements.txt
flask>=2.2,<4
pillow>=10.0
numpy>=1.24
requests>=2.31
gunicorn>=21.2
